requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
httplib2==0.21.0
idna==3.4
multidict==6.0.4
psycopg2-binary==2.9.5
pyparsing==3.0.9
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.14
wikilinks==0.0.4
wikipedia==1.4.0
yarl==1.8.2
